producers/log-producer/requirements.txt
kafka-python==2.0.2
faker==20.1.0

